refactor: split api key resolver
